Reorganize ypplandingview files
