+++ packages/atlas/src/views/global/YppLandingView/YppLandingView.tsx
@@ -17,11 +17,11 @@ import { useYppStore } from '@/providers/ypp/ypp.store'
 import { SentryLogger } from '@/utils/logs'
 
 import { YppAuthorizationModal } from './YppAuthorizationModal'
-import { YppCardsSections } from './YppCardsSections'
-import { YppFooter } from './YppFooter'
-import { YppHero } from './YppHero'
 import { Wrapper } from './YppLandingView.styles'
-import { YppRewardSection } from './YppRewardSection'
+import { YppCardsSections } from './sections/YppCardsSections'
+import { YppFooter } from './sections/YppFooter'
+import { YppHero } from './sections/YppHero'
+import { YppRewardSection } from './sections/YppRewardSection'
 import { useGetYppSyncedChannels } from './useGetYppSyncedChannels'
 
 const SINGUP_DAILY_QUOTA = 500 // 2% of the total daily quota
