+++ packages/atlas/src/views/global/YppLandingView/sections/YppFooter.styles.ts
