+++ packages/atlas/src/views/global/YppLandingView/sections/YppRewardSection.styles.ts
