+++ packages/atlas/src/views/studio/YppDashboard/tabs/YppDashboardMainTab.tsx
@@ -11,8 +11,8 @@ import { atlasConfig } from '@/config'
 import { useClipboard } from '@/hooks/useClipboard'
 import { useSegmentAnalytics } from '@/hooks/useSegmentAnalytics'
 import { useUser } from '@/providers/user/user.hooks'
-import { configYppIconMapper } from '@/views/global/YppLandingView/YppFooter'
-import { calculateReward } from '@/views/global/YppLandingView/YppRewardSection'
+import { configYppIconMapper } from '@/views/global/YppLandingView/sections/YppFooter'
+import { calculateReward } from '@/views/global/YppLandingView/sections/YppRewardSection'
 import { useGetYppSyncedChannels } from '@/views/global/YppLandingView/useGetYppSyncedChannels'
 
 import { RewardsWrapper, StyledBanner, StyledSvgAlertsInformative24, WidgetsWrapper } from './YppDashboardTabs.styles'
